working on a new Game component
